cleanup and move folder
